+++ NEWS.md
@@ -1,15 +1,19 @@
 RedLakesCalc-NEWS
 ================
 <Erik.Leppo@tetratech.com>
-2023-12-05 10:18:57.24616
+2023-12-05 10:22:00.559159
 
 <!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
 
-    #> Last Update: 2023-12-05 10:18:57.273028
+    #> Last Update: 2023-12-05 10:22:00.591012
+
+# RedLakesCalc 0.1.0.9002 (2023-12-05)
+
+- feature: Add Shiny app files from BCGcalc
 
 # RedLakesCalc 0.1.0.9001 (2023-12-05)
 
-- Add base package files
+- docs: Add base package files
   - DESCRIPTION
   - NEWS
   - README